Sleep mode update
fixing some hiccups with sleep mode disabling pairing mode/sending packets in general.

Current bugs to squash:
deep sleep mode is triggered but never entered properly. It just lives in the infinite loop printing "off"
This appears to be a problem with idlemode as the deep sleep works when entered directly via another method

idle mode is not polling the probe pin frequently enough. It will only wake up if the probe is triggered while it is sending a heartbeat.
 We should consider setting up the probe pin as an interrupt when it is in idle mode.